+++ routes/_app.tsx
@@ -1,6 +1,5 @@
 import { AppProps } from '$fresh/server.ts';
 import { Head } from '$fresh/runtime.ts';
-import { Stylesheet } from '../mod.ts';
 
 export default function App({ Component }: AppProps) {
   return (
@@ -8,13 +7,11 @@ export default function App({ Component }: AppProps) {
       <Head>
         <meta charSet='utf-8' />
         <meta name='viewport' content='width=device-width, initial-scale=1.0' />
-        <title>Fresh Garlic - Playground</title>
-        <Stylesheet />
+        <title>Fresh Lunchbox</title>
+        <link rel='stylesheet' href='/fonts/fonts.css' />
       </Head>
       <body class='clr-bg-panel clr-txt-base text--paragraph'>
-        <div class='_screen'>
-          <Component />
-        </div>
+        <Component />
       </body>
     </html>
   );